work around device up/down
When a ratelimited device goes down and back up while ratelimit is
running, the qdisc/class setup for this device is wiped by the kernel
but ratelimit remains unaware of this change. When a new ratelimit
request comes in for a client afterwards, the client setup fails and the
client is given unbridled access.

To work around this, upon failure to ratelimit a new client this commit
tries once to reset the parent device setup.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>